linux-gen: pool: add generic pool module to mempool subsystem
